+++ R/private_check.R
@@ -4,60 +4,95 @@ check_args <- function(...) {
   
   env <- new.env(parent = parent.frame()) 
   
-  env$equal <- function(condition, ...) {
-    message <- paste0(...)
-    if (length(message) == 0) 
-      message <- paste0(as.character(match.call()[2]), " failed.")
-    if (isFALSE(condition)) return(message) else return(TRUE)
+  env$is_true <- function(condition, ...) {
+    
+    if (isFALSE(condition)) {
+      message <- paste0(...)
+      if (length(message) == 0) {
+        message <- paste0(
+          "Argument ", as.character(match.call()[2]), " is ill defined."
+        )
+      }
+      return(message) 
+    } else {
+      return(TRUE)
+    }
   }
   
   env$has_length <- function(arg, l, msg) {
     if (missing(msg)) 
-      msg <- paste0(as.character(match.call()[2]), " not of legnth ", l, ".")
-    env$equal(length(arg) == l, msg)
+      msg <- paste0(
+        "Argument ", as.character(match.call()[2]), " not of length ", l, "."
+      )
+    env$is_true(length(arg) == l, msg)
   }
   
-  env$not <- function(condition, ...) env$equal(!condition, ...)
+  env$not <- function(condition, ...) env$is_true(!condition, ...)
   
   env$one_of <- function(arg, ...) {
     match <- c(...)
     msg <- paste0("'", match, "'")
     if (length(match) == 2) msg <- paste0(msg, collapse = " or ")
-    if (length(match) > 2) msg <- paste0("one of ", paste0(msg, collapse = ", "))
-    env$equal(
+    if (length(match) > 2) {
+      msg <- paste0("one of ", paste0(msg, collapse = ", "))
+    }
+    env$is_true(
       arg %in% match, 
-      as.character(match.call()[2]), " is not ", msg, "."
+      "Argument ", as.character(match.call()[2]), " is not ", msg, "."
     )
   }
   
   env$by_call <- function(arg, fn) {
     args <- formals(fn)
     id <- which(names(args) == as.character(match.call()[2]))
-    if (length(id) == 0) stop("by_call has a no matching arg.")
+    if (length(id) == 0) stop("by_call has no matching arg.")
     match <- eval(args[[id]])
     msg <- paste0("'", match, "'")
     if (length(match) == 2) msg <- paste0(msg, collapse = " or ")
     if (length(match) > 2) msg <- paste0("one of ", paste0(msg, collapse = ", "))
-    env$equal(
-      arg %in% match, as.character(match.call()[2]), " is not ", msg, "."
+    env$is_true(
+      arg %in% match, 
+      "Argument ", as.character(match.call()[2]), " is not ", msg, "."
     )
   }
   
   env$within <- function(arg, lower, upper) {
-    if (!missing(lower))
-      x <- env$not(
-        any(unlist(arg) < lower), as.character(match.call()[3]), " < ", lower
-      ) 
-    if (!missing(upper))
-      x <- env$not(
-        any(unlist(arg) > upper), as.character(match.call()[3]), " > ", upper
-      ) 
-    x
+    env$is_true(
+      arg >= lower && arg <= upper, 
+      "Argument ",
+      as.character(match.call()[2]), 
+      " is not within ", lower, " and ", upper, " (is ", arg, ")"
+    ) 
+  }
+  
+  env$at_least <- function(arg, lower, msg) {
+    
+    if (missing(msg)) {
+      msg <- paste0(
+        "Argument ",
+        as.character(match.call()[2]), 
+        " is not greater or equal to ", lower, " (is ", arg, ")"
+      )
+    }
+    
+    env$is_true(arg >= lower, msg) 
+  }
+  
+  env$at_most <- function(arg, upper, msg) {
+    if(missing(msg)) {
+      msg <- paste0(
+        "Argument ",
+        as.character(match.call()[2]), 
+        " is not less or equal to ", upper, " (is ", arg, ")"
+      )
+    }
+    env$is_true(arg <= upper, msg) 
   }
   
   env$by_class <- function(param, class, ...) {
-    env$equal(
+    env$is_true(
       inherits(param, class), 
+      "Argument ", 
       as.character(match.call()[2]), " is not of class ", class, "."
     )
   }
@@ -65,7 +100,6 @@ check_args <- function(...) {
   out <- vector("list", length(expressions) - 1)
   for(i in 2:length(expressions)) {
     out[i - 1] <- eval(expressions[c(1, i)], envir = env)
-    
   }
   out <- out[sapply(out, function(x) if (!isTRUE(x)) TRUE else FALSE)]
   
@@ -75,6 +109,12 @@ check_args <- function(...) {
   }
 }
 
-# dummy functions
-by_class <- by_call <- equal <- not <- within <- 
-  one_of <- has_length <- function(...) {}
+# 
+
+utils::globalVariables(
+  c(
+    "by_class", "by_call", "not","within", "one_of", "has_length", "is_true",
+    "at_least", "at_most"
+  )
+)
+