+++ inst/shiny/shinyscan.R
@@ -1,5 +1,0 @@
-
-#' @export
-shinyscan <- function() {
-  runApp(list(ui = ui, server = server), launch.browser = TRUE)
-}