+++ misc/sem.R
@@ -1,0 +1,10 @@
+library(scan)
+library(lavaan)
+
+dat <- as.data.frame(exampleAB)
+dat
+
+model <- ' i =~ 1*t1 + 1*t2 + 1*t3 + 1*t4
+           s =~ 0*t1 + 1*t2 + 2*t3 + 3*t4 '
+fit <- growth(model, data=Demo.growth)
+summary(fit)